refactor(phases): extract Local Merging operations to local_merge.rs
- Create src/phases/local_merge.rs with phase5 module contents
- Move execute, load_local_fs, merge_local_files, apply_local_operations
- Update mod.rs to import local_merge and re-export as phase5
- Remove unused Error import from mod.rs
- Update context tracking for next task (extract-orchestrator)
