+++ context/refactor-phases-plan.json
@@ -254,7 +254,7 @@
     {
       "id": "extract-phase5-local-merge",
       "name": "Extract phase5 core (Local Merging) to local_merge.rs",
-      "status": "pending",
+      "status": "complete",
       "priority": 5,
       "blocked_by": ["extract-yaml-merge", "extract-json-merge", "extract-toml-merge", "extract-ini-merge", "extract-markdown-merge"],
       "output_file": "src/phases/local_merge.rs",
